Use `kind_arg_list` rather than `letter_spec_list` in `IMPLICIT`
By substituting `kind_arg_list` for `letter_spec_list` in the
`implicit_spec` rule, we can accept the full syntax of the
kind-selector in implicit statements. This involves a number of changes:

    1. Remove `kind` handling from the ImplicitSpec AST definition.
    2. Remove `letter_spec` related rules from grammar.
    3. Add a `kind_arg_list` -> `letter_spec_list` conversion function
       in `semantic.h`

Additionally, it was discovered that `WARN_COMPLEXSTAR` was producing
incorrect warnings due to kind/star disagreement.

Because of the structure of the `implicit_spec` grammar, it is not
possible to point only to the _declaration-type-spec_ portion of the
_implicit-spec_ when issuing a warning about old Sun-style
declarations (e.g. `REAL*8`). The entire `implicit_spec` was being
underlined, which was a little confusing.  Instead, we are now
pointing to just the `*` location, which, while not as good as
underlining all of `REAL*8`, at least directs the reader to the
suspect construct.